split schema.hs into multiple modules
