+++ server/src-lib/Hasura/GraphQL/Schema.hs
@@ -10,31 +10,38 @@ module Hasura.GraphQL.Schema
   , RelationInfoMap
   , isAggFld
   , qualObjectToName
-  -- Schema stitching related
-  , checkSchemaConflicts
-  , checkConflictingNode
-  , emptyGCtx
-  , mergeMaybeMaps
   , ppGCtx
+
+  , checkConflictingNode
+  , checkSchemaConflicts
   ) where
 
 
-import           Data.Maybe                     (maybeToList)
+import qualified Data.HashMap.Strict                   as Map
+import qualified Data.HashSet                          as Set
+import qualified Data.Sequence                         as Seq
 
-import qualified Data.HashMap.Strict            as Map
-import qualified Data.HashSet                   as Set
-import qualified Data.Sequence                  as Seq
-
-import qualified Data.Text                      as T
-import qualified Language.GraphQL.Draft.Syntax  as G
+import qualified Data.Text                             as T
+import qualified Language.GraphQL.Draft.Syntax         as G
 
 import           Hasura.GraphQL.Context
 import           Hasura.GraphQL.Resolve.Types
 import           Hasura.GraphQL.Validate.Types
 import           Hasura.Prelude
-import           Hasura.RQL.DML.Internal        (mkAdminRolePermInfo)
+import           Hasura.RQL.DML.Internal               (mkAdminRolePermInfo)
 import           Hasura.RQL.Types
 import           Hasura.SQL.Types
+
+import           Hasura.GraphQL.Schema.BoolExp
+import           Hasura.GraphQL.Schema.Common
+import           Hasura.GraphQL.Schema.Function
+import           Hasura.GraphQL.Schema.Mutation.Common
+import           Hasura.GraphQL.Schema.Mutation.Delete
+import           Hasura.GraphQL.Schema.Mutation.Insert
+import           Hasura.GraphQL.Schema.Mutation.Update
+import           Hasura.GraphQL.Schema.OrderBy
+import           Hasura.GraphQL.Schema.Select
+import           Hasura.GraphQL.Schema.Stitch
 
 getInsPerm :: TableInfo -> RoleName -> Maybe InsPermInfo
 getInsPerm tabInfo role
@@ -49,11 +56,6 @@ getTabInfo
 getTabInfo tc t =
   onNothing (Map.lookup t tc) $
      throw500 $ "table not found: " <>> t
-
-type SelField = Either PGColInfo (RelInfo, Bool, AnnBoolExpPartialSQL, Maybe Int, Bool)
-
-qualObjectToName :: (ToTxt a) => QualifiedObject a -> G.Name
-qualObjectToName = G.Name . snakeCaseQualObject
 
 isValidObjectName :: (ToTxt a) => QualifiedObject a -> Bool
 isValidObjectName = isValidName . qualObjectToName
@@ -107,1108 +109,6 @@ compAggOps = ["max", "min"]
 
 isAggFld :: G.Name -> Bool
 isAggFld = flip elem (numAggOps <> compAggOps)
-
-mkColName :: PGCol -> G.Name
-mkColName (PGCol n) = G.Name n
-
-mkRelName :: RelName -> G.Name
-mkRelName rn = G.Name $ relNameToTxt rn
-
-mkAggRelName :: RelName -> G.Name
-mkAggRelName rn = G.Name $ relNameToTxt rn <> "_aggregate"
-
-mkBoolExpName :: QualifiedTable -> G.Name
-mkBoolExpName tn =
-  qualObjectToName tn <> "_bool_exp"
-
-mkBoolExpTy :: QualifiedTable -> G.NamedType
-mkBoolExpTy =
-  G.NamedType . mkBoolExpName
-
-mkFuncArgsName :: QualifiedFunction -> G.Name
-mkFuncArgsName fn =
-  qualObjectToName fn <> "_args"
-
-mkFuncArgsTy :: QualifiedFunction -> G.NamedType
-mkFuncArgsTy =
-  G.NamedType . mkFuncArgsName
-
-mkTableTy :: QualifiedTable -> G.NamedType
-mkTableTy =
-  G.NamedType . qualObjectToName
-
-mkTableAggTy :: QualifiedTable -> G.NamedType
-mkTableAggTy tn =
-  G.NamedType $ qualObjectToName tn <> "_aggregate"
-
-mkTableAggFldsTy :: QualifiedTable -> G.NamedType
-mkTableAggFldsTy tn =
-  G.NamedType $ qualObjectToName tn <> "_aggregate_fields"
-
-mkTableColAggFldsTy :: G.Name -> QualifiedTable -> G.NamedType
-mkTableColAggFldsTy op tn =
-  G.NamedType $ qualObjectToName tn <> "_" <> op <> "_fields"
-
-mkTableByPkName :: QualifiedTable -> G.Name
-mkTableByPkName tn = qualObjectToName tn <> "_by_pk"
-
--- Support argument params for PG columns
-mkPGColParams :: PGColType -> ParamMap
-mkPGColParams = \case
-  PGJSONB -> jsonParams
-  PGJSON  -> jsonParams
-  _       -> Map.empty
-  where
-    pathDesc = "JSON select path"
-    jsonParams = Map.fromList
-      [ (G.Name "path", InpValInfo (Just pathDesc) "path" Nothing $ G.toGT $ mkScalarTy PGText)
-      ]
-
-mkPGColFld :: PGColInfo -> ObjFldInfo
-mkPGColFld (PGColInfo colName colTy isNullable) =
-  mkHsraObjFldInfo Nothing n (mkPGColParams colTy) ty
-  where
-    n  = G.Name $ getPGColTxt colName
-    ty = bool notNullTy nullTy isNullable
-    scalarTy = mkScalarTy colTy
-    notNullTy = G.toGT $ G.toNT scalarTy
-    nullTy = G.toGT scalarTy
-
--- where: table_bool_exp
--- limit: Int
--- offset: Int
--- distinct_on: [table_select_column!]
-mkSelArgs :: QualifiedTable -> [InpValInfo]
-mkSelArgs tn =
-  [ InpValInfo (Just whereDesc) "where" Nothing $ G.toGT $ mkBoolExpTy tn
-  , InpValInfo (Just limitDesc) "limit" Nothing $ G.toGT $ mkScalarTy PGInteger
-  , InpValInfo (Just offsetDesc) "offset" Nothing $ G.toGT $ mkScalarTy PGInteger
-  , InpValInfo (Just orderByDesc) "order_by" Nothing $ G.toGT $ G.toLT $ G.toNT $
-    mkOrdByTy tn
-  , InpValInfo (Just distinctDesc) "distinct_on" Nothing $ G.toGT $ G.toLT $
-    G.toNT $ mkSelColumnInpTy tn
-  ]
-  where
-    whereDesc   = "filter the rows returned"
-    limitDesc   = "limit the nuber of rows returned"
-    offsetDesc  = "skip the first n rows. Use only with order_by"
-    orderByDesc = "sort the rows by one or more columns"
-    distinctDesc = "distinct select on columns"
-
--- fromInpValL :: [InpValInfo] -> Map.HashMap G.Name InpValInfo
--- fromInpValL = mapFromL _iviName
-
-{-
-
-array_relationship(
-  where: remote_table_bool_exp
-  limit: Int
-  offset: Int
-):  [remote_table!]!
-array_relationship_aggregate(
-  where: remote_table_bool_exp
-  limit: Int
-  offset: Int
-):  remote_table_aggregate!
-object_relationship: remote_table
-
--}
-mkRelFld
-  :: Bool
-  -> RelInfo
-  -> Bool
-  -> [ObjFldInfo]
-mkRelFld allowAgg (RelInfo rn rTy _ remTab isManual) isNullable = case rTy of
-  ArrRel -> bool [arrRelFld] [arrRelFld, aggArrRelFld] allowAgg
-  ObjRel -> [objRelFld]
-  where
-    objRelFld = mkHsraObjFldInfo (Just "An object relationship")
-      (mkRelName rn) Map.empty objRelTy
-    objRelTy = bool (G.toGT $ G.toNT relTabTy) (G.toGT relTabTy) isObjRelNullable
-    isObjRelNullable = isManual || isNullable
-    relTabTy = mkTableTy remTab
-
-    arrRelFld =
-      mkHsraObjFldInfo (Just "An array relationship") (mkRelName rn)
-      (fromInpValL $ mkSelArgs remTab) arrRelTy
-    arrRelTy = G.toGT $ G.toNT $ G.toLT $ G.toNT $ mkTableTy remTab
-    aggArrRelFld = mkHsraObjFldInfo (Just "An aggregated array relationship")
-      (mkAggRelName rn) (fromInpValL $ mkSelArgs remTab) $
-      G.toGT $ G.toNT $ mkTableAggTy remTab
-
-{-
-type table {
-  col1: colty1
-  .
-  .
-  rel1: relty1
-}
--}
-mkTableObj
-  :: QualifiedTable
-  -> [SelField]
-  -> ObjTyInfo
-mkTableObj tn allowedFlds =
-  mkObjTyInfo (Just desc) (mkTableTy tn) Set.empty (mapFromL _fiName flds) TLHasuraType
-  where
-    flds = concatMap (either (pure . mkPGColFld) mkRelFld') allowedFlds
-    mkRelFld' (relInfo, allowAgg, _, _, isNullable) =
-      mkRelFld allowAgg relInfo isNullable
-    desc = G.Description $ "columns and relationships of " <>> tn
-
-{-
-type table_aggregate {
-  agg: table_aggregate_fields
-  nodes: [table!]!
-}
--}
-mkTableAggObj
-  :: QualifiedTable -> ObjTyInfo
-mkTableAggObj tn =
-  mkHsraObjTyInfo (Just desc) (mkTableAggTy tn) Set.empty $ mapFromL _fiName
-  [aggFld, nodesFld]
-  where
-    desc = G.Description $
-      "aggregated selection of " <>> tn
-
-    aggFld = mkHsraObjFldInfo Nothing "aggregate" Map.empty $ G.toGT $
-             mkTableAggFldsTy tn
-    nodesFld = mkHsraObjFldInfo Nothing "nodes" Map.empty $ G.toGT $
-               G.toNT $ G.toLT $ G.toNT $ mkTableTy tn
-
-{-
-type table_aggregate_fields{
-  count: Int
-  sum: table_sum_fields
-  avg: table_avg_fields
-  stddev: table_stddev_fields
-  stddev_pop: table_stddev_pop_fields
-  variance: table_variance_fields
-  var_pop: table_var_pop_fields
-  max: table_max_fields
-  min: table_min_fields
-}
--}
-mkTableAggFldsObj
-  :: QualifiedTable -> [PGCol] -> [PGCol] -> ObjTyInfo
-mkTableAggFldsObj tn numCols compCols =
-  mkHsraObjTyInfo (Just desc) (mkTableAggFldsTy tn) Set.empty $ mapFromL _fiName $
-  countFld : (numFlds <> compFlds)
-  where
-    desc = G.Description $
-      "aggregate fields of " <>> tn
-
-    countFld = mkHsraObjFldInfo Nothing "count" countParams $ G.toGT $
-               mkScalarTy PGInteger
-
-    countParams = fromInpValL [countColInpVal, distinctInpVal]
-
-    countColInpVal = InpValInfo Nothing "columns" Nothing $ G.toGT $
-                     G.toLT $ G.toNT $ mkSelColumnInpTy tn
-    distinctInpVal = InpValInfo Nothing "distinct" Nothing $ G.toGT $
-                     mkScalarTy PGBoolean
-
-    numFlds = bool (map mkColOpFld numAggOps) [] $ null numCols
-    compFlds = bool (map mkColOpFld compAggOps) [] $ null compCols
-
-    mkColOpFld op = mkHsraObjFldInfo Nothing op Map.empty $ G.toGT $
-                    mkTableColAggFldsTy op tn
-
-{-
-type table_<agg-op>_fields{
-   num_col: Int
-   .        .
-   .        .
-}
--}
-mkTableColAggFldsObj
-  :: QualifiedTable
-  -> G.Name
-  -> (PGColType -> G.NamedType)
-  -> [PGColInfo]
-  -> ObjTyInfo
-mkTableColAggFldsObj tn op f cols =
-  mkHsraObjTyInfo (Just desc) (mkTableColAggFldsTy op tn) Set.empty $ mapFromL _fiName $
-  map mkColObjFld cols
-  where
-    desc = G.Description $ "aggregate " <> G.unName op <> " on columns"
-
-    mkColObjFld c = mkHsraObjFldInfo Nothing (G.Name $ getPGColTxt $ pgiName c)
-                    Map.empty $ G.toGT $ f $ pgiType c
-
-{-
-
-table(
-  where: table_bool_exp
-  limit: Int
-  offset: Int
-):  [table!]!
-
--}
-mkSelFld
-  :: QualifiedTable
-  -> ObjFldInfo
-mkSelFld tn =
-  mkHsraObjFldInfo (Just desc) fldName args ty
-  where
-    desc    = G.Description $ "fetch data from the table: " <>> tn
-    fldName = qualObjectToName tn
-    args    = fromInpValL $ mkSelArgs tn
-    ty      = G.toGT $ G.toNT $ G.toLT $ G.toNT $ mkTableTy tn
-
-{-
-table_by_pk(
-  col1: value1!,
-  .     .
-  .     .
-  coln: valuen!
-): table
--}
-mkSelFldPKey
-  :: QualifiedTable -> [PGColInfo]
-  -> ObjFldInfo
-mkSelFldPKey tn cols =
-  mkHsraObjFldInfo (Just desc) fldName args ty
-  where
-    desc = G.Description $ "fetch data from the table: " <> tn
-           <<> " using primary key columns"
-    fldName = mkTableByPkName tn
-    args = fromInpValL $ map colInpVal cols
-    ty = G.toGT $ mkTableTy tn
-    colInpVal (PGColInfo n typ _) =
-      InpValInfo Nothing (mkColName n) Nothing $ G.toGT $ G.toNT $ mkScalarTy typ
-
-{-
-
-table_aggregate(
-  where: table_bool_exp
-  limit: Int
-  offset: Int
-): table_aggregate!
-
--}
-mkAggSelFld
-  :: QualifiedTable
-  -> ObjFldInfo
-mkAggSelFld tn =
-  mkHsraObjFldInfo (Just desc) fldName args ty
-  where
-    desc = G.Description $ "fetch aggregated fields from the table: "
-           <>> tn
-    fldName = qualObjectToName tn <> "_aggregate"
-    args = fromInpValL $ mkSelArgs tn
-    ty = G.toGT $ G.toNT $ mkTableAggTy tn
-
-{-
-
-function(
-  args: function_args
-  where: table_bool_exp
-  limit: Int
-  offset: Int
-): [table!]!
-
--}
-
-mkFuncArgs :: FunctionInfo -> ParamMap
-mkFuncArgs funInfo =
-  fromInpValL $ funcInpArgs <> mkSelArgs retTable
-  where
-    funcName = fiName funInfo
-    funcArgs = fiInputArgs funInfo
-    retTable = fiReturnType funInfo
-
-    funcArgDesc = G.Description $ "input parameters for function " <>> funcName
-    funcInpArg = InpValInfo (Just funcArgDesc) "args" Nothing $ G.toGT $ G.toNT $
-                 mkFuncArgsTy funcName
-    funcInpArgs = bool [funcInpArg] [] $ null funcArgs
-
-mkFuncQueryFld
-  :: FunctionInfo -> ObjFldInfo
-mkFuncQueryFld funInfo =
-  mkHsraObjFldInfo (Just desc) fldName (mkFuncArgs funInfo) ty
-  where
-    retTable = fiReturnType funInfo
-    funcName = fiName funInfo
-
-    desc = G.Description $ "execute function " <> funcName
-           <<> " which returns " <>> retTable
-    fldName = qualObjectToName funcName
-
-    ty      = G.toGT $ G.toNT $ G.toLT $ G.toNT $ mkTableTy retTable
-
-{-
-
-function_aggregate(
-  args: function_args
-  where: table_bool_exp
-  limit: Int
-  offset: Int
-): table_aggregate!
-
--}
-
-mkFuncAggQueryFld
-  :: FunctionInfo -> ObjFldInfo
-mkFuncAggQueryFld funInfo =
-  mkHsraObjFldInfo (Just desc) fldName (mkFuncArgs funInfo) ty
-  where
-    funcName = fiName funInfo
-    retTable = fiReturnType funInfo
-
-    desc = G.Description $ "execute function " <> funcName
-           <<> " and query aggregates on result of table type "
-           <>> retTable
-
-    fldName = qualObjectToName funcName <> "_aggregate"
-
-    ty = G.toGT $ G.toNT $ mkTableAggTy retTable
-
-
--- table_mutation_response
-mkMutRespTy :: QualifiedTable -> G.NamedType
-mkMutRespTy tn =
-  G.NamedType $ qualObjectToName tn <> "_mutation_response"
-
-{-
-type table_mutation_response {
-  affected_rows: Int!
-  returning: [table!]!
-}
--}
-mkMutRespObj
-  :: QualifiedTable
-  -> Bool -- is sel perm defined
-  -> ObjTyInfo
-mkMutRespObj tn sel =
-  mkHsraObjTyInfo (Just objDesc) (mkMutRespTy tn) Set.empty $ mapFromL _fiName
-    $ affectedRowsFld : bool [] [returningFld] sel
-  where
-    objDesc = G.Description $
-      "response of any mutation on the table " <>> tn
-    affectedRowsFld =
-      mkHsraObjFldInfo (Just desc) "affected_rows" Map.empty $
-        G.toGT $ G.toNT $ mkScalarTy PGInteger
-      where
-        desc = "number of affected rows by the mutation"
-    returningFld =
-      mkHsraObjFldInfo (Just desc) "returning" Map.empty $
-        G.toGT $ G.toNT $ G.toLT $ G.toNT $ mkTableTy tn
-      where
-        desc = "data of the affected rows by the mutation"
-
--- table_bool_exp
-mkBoolExpInp
-  :: QualifiedTable
-  -- the fields that are allowed
-  -> [SelField]
-  -> InpObjTyInfo
-mkBoolExpInp tn fields =
-  mkHsraInpTyInfo (Just desc) boolExpTy $ Map.fromList
-    [(_iviName inpVal, inpVal) | inpVal <- inpValues]
-  where
-    desc = G.Description $
-      "Boolean expression to filter rows from the table " <> tn <<>
-      ". All fields are combined with a logical 'AND'."
-
-    -- the type of this boolean expression
-    boolExpTy = mkBoolExpTy tn
-
-    -- all the fields of this input object
-    inpValues = combinators <> map mkFldExpInp fields
-
-    mk n ty = InpValInfo Nothing n Nothing $ G.toGT ty
-
-    boolExpListTy = G.toLT boolExpTy
-
-    combinators =
-      [ mk "_not" boolExpTy
-      , mk "_and" boolExpListTy
-      , mk "_or"  boolExpListTy
-      ]
-
-    mkFldExpInp = \case
-      Left (PGColInfo colName colTy _) ->
-        mk (mkColName colName) (mkCompExpTy colTy)
-      Right (RelInfo relName _ _ remTab _, _, _, _, _) ->
-        mk (mkRelName relName) (mkBoolExpTy remTab)
-
-mkPGColInp :: PGColInfo -> InpValInfo
-mkPGColInp (PGColInfo colName colTy _) =
-  InpValInfo Nothing (G.Name $ getPGColTxt colName) Nothing $
-  G.toGT $ mkScalarTy colTy
-
-{-
-input function_args {
-  arg1: arg-type1!
-  .     .
-  .     .
-  argn: arg-typen!
-}
--}
-
-procFuncArgs
-  :: Seq.Seq FunctionArg
-  -> (PGColType -> Text -> a) -> [a]
-procFuncArgs argSeq f =
-  fst $ foldl mkItem ([], 1::Int) argSeq
-  where
-    mkItem (items, argNo) (FunctionArg nameM ty) =
-      case nameM of
-        Just argName ->
-          let argT = getFuncArgNameTxt argName
-          in (items <> pure (f ty argT), argNo)
-        Nothing ->
-          let argT = "arg_" <> T.pack (show argNo)
-          in (items <> pure (f ty argT), argNo + 1)
-
-mkFuncArgsInp :: FunctionInfo -> Maybe InpObjTyInfo
-mkFuncArgsInp funcInfo =
-  bool (Just inpObj) Nothing $ null funcArgs
-  where
-    funcName = fiName funcInfo
-    funcArgs = fiInputArgs funcInfo
-    funcArgsTy = mkFuncArgsTy funcName
-
-    inpObj = mkHsraInpTyInfo Nothing funcArgsTy $
-             fromInpValL argInps
-
-    argInps = procFuncArgs funcArgs mkInpVal
-
-    mkInpVal ty t =
-      InpValInfo Nothing (G.Name t) Nothing $
-      G.toGT $ mkScalarTy ty
-
--- table_set_input
-mkUpdSetTy :: QualifiedTable -> G.NamedType
-mkUpdSetTy tn =
-  G.NamedType $ qualObjectToName tn <> "_set_input"
-
-{-
-input table_set_input {
-  col1: colty1
-  .
-  .
-  coln: coltyn
-}
--}
-mkUpdSetInp
-  :: QualifiedTable -> [PGColInfo] -> InpObjTyInfo
-mkUpdSetInp tn cols  =
-  mkHsraInpTyInfo (Just desc) (mkUpdSetTy tn) $
-    fromInpValL $ map mkPGColInp cols
-  where
-    desc = G.Description $
-      "input type for updating data in table " <>> tn
-
--- table_inc_input
-mkUpdIncTy :: QualifiedTable -> G.NamedType
-mkUpdIncTy tn =
-  G.NamedType $ qualObjectToName tn <> "_inc_input"
-
-{-
-input table_inc_input {
-  integer-col1: int
-  .
-  .
-  integer-coln: int
-}
--}
-
-mkUpdIncInp
-  :: QualifiedTable -> Maybe [PGColInfo] -> Maybe InpObjTyInfo
-mkUpdIncInp tn = maybe Nothing mkType
-  where
-    mkType cols = let intCols = onlyIntCols cols
-                      incObjTy =
-                        mkHsraInpTyInfo (Just desc) (mkUpdIncTy tn) $
-                          fromInpValL $ map mkPGColInp intCols
-                  in bool (Just incObjTy) Nothing $ null intCols
-    desc = G.Description $
-      "input type for incrementing integer columne in table " <>> tn
-
--- table_<json-op>_input
-mkJSONOpTy :: QualifiedTable -> G.Name -> G.NamedType
-mkJSONOpTy tn op =
-  G.NamedType $ qualObjectToName tn <> op <> "_input"
-
--- json ops are _concat, _delete_key, _delete_elem, _delete_at_path
-{-
-input table_concat_input {
-  jsonb-col1: json
-  .
-  .
-  jsonb-coln: json
-}
--}
-
-{-
-input table_delete_key_input {
-  jsonb-col1: string
-  .
-  .
-  jsonb-coln: string
-}
--}
-
-{-
-input table_delete_elem_input {
-  jsonb-col1: int
-  .
-  .
-  jsonb-coln: int
-}
--}
-
-{-
-input table_delete_at_path_input {
-  jsonb-col1: [string]
-  .
-  .
-  jsonb-coln: [string]
-}
--}
-
--- jsonb operators and descriptions
-prependOp :: G.Name
-prependOp = "_prepend"
-
-prependDesc :: G.Description
-prependDesc = "prepend existing jsonb value of filtered columns with new jsonb value"
-
-appendOp :: G.Name
-appendOp = "_append"
-
-appendDesc :: G.Description
-appendDesc = "append existing jsonb value of filtered columns with new jsonb value"
-
-deleteKeyOp :: G.Name
-deleteKeyOp = "_delete_key"
-
-deleteKeyDesc :: G.Description
-deleteKeyDesc = "delete key/value pair or string element."
-                <> " key/value pairs are matched based on their key value"
-
-deleteElemOp :: G.Name
-deleteElemOp = "_delete_elem"
-
-deleteElemDesc :: G.Description
-deleteElemDesc = "delete the array element with specified index (negative integers count from the end)."
-                 <> " throws an error if top level container is not an array"
-
-deleteAtPathOp :: G.Name
-deleteAtPathOp = "_delete_at_path"
-
-deleteAtPathDesc :: G.Description
-deleteAtPathDesc = "delete the field or element with specified path"
-                   <> " (for JSON arrays, negative integers count from the end)"
-
-mkUpdJSONOpInp
-  :: QualifiedTable -> [PGColInfo] -> [InpObjTyInfo]
-mkUpdJSONOpInp tn cols = bool inpObjs [] $ null jsonbCols
-  where
-    jsonbCols = onlyJSONBCols cols
-    jsonbColNames = map pgiName jsonbCols
-
-    inpObjs = [ prependInpObj, appendInpObj, deleteKeyInpObj
-              , deleteElemInpObj, deleteAtPathInpObj
-              ]
-
-    appendInpObj =
-      mkHsraInpTyInfo (Just appendDesc) (mkJSONOpTy tn appendOp) $
-      fromInpValL $ map mkPGColInp jsonbCols
-
-    prependInpObj =
-      mkHsraInpTyInfo (Just prependDesc) (mkJSONOpTy tn prependOp) $
-      fromInpValL $ map mkPGColInp jsonbCols
-
-    deleteKeyInpObj =
-      mkHsraInpTyInfo (Just deleteKeyDesc) (mkJSONOpTy tn deleteKeyOp) $
-      fromInpValL $ map deleteKeyInpVal jsonbColNames
-    deleteKeyInpVal c = InpValInfo Nothing (G.Name $ getPGColTxt c) Nothing $
-      G.toGT $ G.NamedType "String"
-
-    deleteElemInpObj =
-      mkHsraInpTyInfo (Just deleteElemDesc) (mkJSONOpTy tn deleteElemOp) $
-      fromInpValL $ map deleteElemInpVal jsonbColNames
-    deleteElemInpVal c = InpValInfo Nothing (G.Name $ getPGColTxt c) Nothing $
-      G.toGT $ G.NamedType "Int"
-
-    deleteAtPathInpObj =
-      mkHsraInpTyInfo (Just deleteAtPathDesc) (mkJSONOpTy tn deleteAtPathOp) $
-      fromInpValL $ map deleteAtPathInpVal jsonbColNames
-    deleteAtPathInpVal c = InpValInfo Nothing (G.Name $ getPGColTxt c) Nothing $
-      G.toGT $ G.toLT $ G.NamedType "String"
-
-{-
-
-update_table(
-  where : table_bool_exp!
-  _set  : table_set_input
-  _inc  : table_inc_input
-  _concat: table_concat_input
-  _delete_key: table_delete_key_input
-  _delete_elem: table_delete_elem_input
-  _delete_path_at: table_delete_path_at_input
-): table_mutation_response
-
--}
-
-mkIncInpVal :: QualifiedTable -> [PGColInfo] -> Maybe InpValInfo
-mkIncInpVal tn cols = bool (Just incArg) Nothing $ null intCols
-  where
-    intCols = onlyIntCols cols
-    incArgDesc = "increments the integer columns with given value of the filtered values"
-    incArg =
-      InpValInfo (Just incArgDesc) "_inc" Nothing $ G.toGT $ mkUpdIncTy tn
-
-mkJSONOpInpVals :: QualifiedTable -> [PGColInfo] -> [InpValInfo]
-mkJSONOpInpVals tn cols = bool jsonbOpArgs [] $ null jsonbCols
-  where
-    jsonbCols = onlyJSONBCols cols
-    jsonbOpArgs = [appendArg, prependArg, deleteKeyArg, deleteElemArg, deleteAtPathArg]
-
-    appendArg =
-      InpValInfo (Just appendDesc) appendOp Nothing $ G.toGT $ mkJSONOpTy tn appendOp
-
-    prependArg =
-      InpValInfo (Just prependDesc) prependOp Nothing $ G.toGT $ mkJSONOpTy tn prependOp
-
-    deleteKeyArg =
-      InpValInfo (Just deleteKeyDesc) deleteKeyOp Nothing $
-      G.toGT $ mkJSONOpTy tn deleteKeyOp
-
-    deleteElemArg =
-      InpValInfo (Just deleteElemDesc) deleteElemOp Nothing $
-      G.toGT $ mkJSONOpTy tn deleteElemOp
-
-    deleteAtPathArg =
-      InpValInfo (Just deleteAtPathDesc) deleteAtPathOp Nothing $
-      G.toGT $ mkJSONOpTy tn deleteAtPathOp
-
-mkUpdMutFld
-  :: QualifiedTable -> [PGColInfo] -> ObjFldInfo
-mkUpdMutFld tn cols =
-  mkHsraObjFldInfo (Just desc) fldName (fromInpValL inputValues) $
-    G.toGT $ mkMutRespTy tn
-  where
-    inputValues = [filterArg, setArg] <> incArg
-                  <> mkJSONOpInpVals tn cols
-    desc = G.Description $ "update data of the table: " <>> tn
-
-    fldName = "update_" <> qualObjectToName tn
-
-    filterArgDesc = "filter the rows which have to be updated"
-    filterArg =
-      InpValInfo (Just filterArgDesc) "where" Nothing $ G.toGT $
-      G.toNT $ mkBoolExpTy tn
-
-    setArgDesc = "sets the columns of the filtered rows to the given values"
-    setArg =
-      InpValInfo (Just setArgDesc) "_set" Nothing $ G.toGT $ mkUpdSetTy tn
-
-    incArg = maybeToList $ mkIncInpVal tn cols
-
-{-
-
-delete_table(
-  where : table_bool_exp!
-): table_mutation_response
-
--}
-
-mkDelMutFld
-  :: QualifiedTable -> ObjFldInfo
-mkDelMutFld tn =
-  mkHsraObjFldInfo (Just desc) fldName (fromInpValL [filterArg]) $
-    G.toGT $ mkMutRespTy tn
-  where
-    desc = G.Description $ "delete data from the table: " <>> tn
-
-    fldName = "delete_" <> qualObjectToName tn
-
-    filterArgDesc = "filter the rows which have to be deleted"
-    filterArg =
-      InpValInfo (Just filterArgDesc) "where" Nothing $ G.toGT $
-      G.toNT $ mkBoolExpTy tn
-
--- table_insert_input
-mkInsInpTy :: QualifiedTable -> G.NamedType
-mkInsInpTy tn =
-  G.NamedType $ qualObjectToName tn <> "_insert_input"
-
--- table_obj_rel_insert_input
-mkObjInsInpTy :: QualifiedTable -> G.NamedType
-mkObjInsInpTy tn =
-  G.NamedType $ qualObjectToName tn <> "_obj_rel_insert_input"
-
--- table_arr_rel_insert_input
-mkArrInsInpTy :: QualifiedTable -> G.NamedType
-mkArrInsInpTy tn =
-  G.NamedType $ qualObjectToName tn <> "_arr_rel_insert_input"
-
-
--- table_on_conflict
-mkOnConflictInpTy :: QualifiedTable -> G.NamedType
-mkOnConflictInpTy tn =
-  G.NamedType $ qualObjectToName tn <> "_on_conflict"
-
--- table_constraint
-mkConstraintInpTy :: QualifiedTable -> G.NamedType
-mkConstraintInpTy tn =
-  G.NamedType $ qualObjectToName tn <> "_constraint"
-
--- conflict_action
-conflictActionTy :: G.NamedType
-conflictActionTy = G.NamedType "conflict_action"
-
--- table_update_column
-mkUpdColumnInpTy :: QualifiedTable -> G.NamedType
-mkUpdColumnInpTy tn =
-  G.NamedType $ qualObjectToName tn <> "_update_column"
-
---table_select_column
-mkSelColumnInpTy :: QualifiedTable -> G.NamedType
-mkSelColumnInpTy tn =
-  G.NamedType $ qualObjectToName tn <> "_select_column"
-
-{-
-input table_obj_rel_insert_input {
-  data: table_insert_input!
-  on_conflict: table_on_conflict
-}
-
--}
-
-{-
-input table_arr_rel_insert_input {
-  data: [table_insert_input!]!
-  on_conflict: table_on_conflict
-}
-
--}
-
-mkRelInsInps
-  :: QualifiedTable -> Bool -> [InpObjTyInfo]
-mkRelInsInps tn upsertAllowed = [objRelInsInp, arrRelInsInp]
-  where
-    onConflictInpVal =
-      InpValInfo Nothing "on_conflict" Nothing $ G.toGT $ mkOnConflictInpTy tn
-
-    onConflictInp = bool [] [onConflictInpVal] upsertAllowed
-
-    objRelDesc = G.Description $
-      "input type for inserting object relation for remote table " <>> tn
-
-    objRelDataInp = InpValInfo Nothing "data" Nothing $ G.toGT $
-                    G.toNT $ mkInsInpTy tn
-    objRelInsInp = mkHsraInpTyInfo (Just objRelDesc) (mkObjInsInpTy tn)
-                   $ fromInpValL $ objRelDataInp : onConflictInp
-
-    arrRelDesc = G.Description $
-      "input type for inserting array relation for remote table " <>> tn
-
-    arrRelDataInp = InpValInfo Nothing "data" Nothing $ G.toGT $
-                    G.toNT $ G.toLT $ G.toNT $ mkInsInpTy tn
-    arrRelInsInp = mkHsraInpTyInfo (Just arrRelDesc) (mkArrInsInpTy tn)
-                   $ fromInpValL $ arrRelDataInp : onConflictInp
-
-{-
-
-input table_insert_input {
-  col1: colty1
-  .
-  .
-  coln: coltyn
-}
-
--}
-
-mkInsInp
-  :: QualifiedTable -> [PGColInfo] -> RelationInfoMap -> InpObjTyInfo
-mkInsInp tn insCols relInfoMap =
-  mkHsraInpTyInfo (Just desc) (mkInsInpTy tn) $ fromInpValL $
-  map mkPGColInp insCols <> relInps
-  where
-    desc = G.Description $
-      "input type for inserting data into table " <>> tn
-
-    relInps = flip map (Map.toList relInfoMap) $
-      \(relName, relInfo) ->
-         let remoteQT = riRTable relInfo
-             tyMaker = case riType relInfo of
-               ObjRel -> mkObjInsInpTy
-               ArrRel -> mkArrInsInpTy
-         in  InpValInfo Nothing (mkRelName relName) Nothing $
-               G.toGT $ tyMaker remoteQT
-
-
-{-
-
-input table_on_conflict {
-  action: conflict_action
-  constraint: table_constraint!
-  update_columns: [table_column!]
-}
-
--}
-
-mkOnConflictInp :: QualifiedTable -> InpObjTyInfo
-mkOnConflictInp tn =
-  mkHsraInpTyInfo (Just desc) (mkOnConflictInpTy tn) $ fromInpValL
-  [constraintInpVal, updateColumnsInpVal]
-  where
-    desc = G.Description $
-      "on conflict condition type for table " <>> tn
-
-    constraintInpVal = InpValInfo Nothing (G.Name "constraint") Nothing $
-      G.toGT $ G.toNT $ mkConstraintInpTy tn
-
-    updateColumnsInpVal = InpValInfo Nothing (G.Name "update_columns") Nothing $
-      G.toGT $ G.toNT $ G.toLT $ G.toNT $ mkUpdColumnInpTy tn
-{-
-
-insert_table(
-  objects: [table_insert_input!]!
-  on_conflict: table_on_conflict
-  ): table_mutation_response!
--}
-
-mkInsMutFld
-  :: QualifiedTable -> Bool -> ObjFldInfo
-mkInsMutFld tn isUpsertable =
-  mkHsraObjFldInfo (Just desc) fldName (fromInpValL inputVals) $
-    G.toGT $ mkMutRespTy tn
-  where
-    inputVals = catMaybes [Just objectsArg , onConflictInpVal]
-    desc = G.Description $
-      "insert data into the table: " <>> tn
-
-    fldName = "insert_" <> qualObjectToName tn
-
-    objsArgDesc = "the rows to be inserted"
-    objectsArg =
-      InpValInfo (Just objsArgDesc) "objects" Nothing $ G.toGT $
-      G.toNT $ G.toLT $ G.toNT $ mkInsInpTy tn
-
-    onConflictInpVal = bool Nothing (Just onConflictArg) isUpsertable
-
-    onConflictDesc = "on conflict condition"
-    onConflictArg =
-      InpValInfo (Just onConflictDesc) "on_conflict" Nothing $ G.toGT $ mkOnConflictInpTy tn
-
-mkConstriantTy :: QualifiedTable -> [ConstraintName] -> EnumTyInfo
-mkConstriantTy tn cons = enumTyInfo
-  where
-    enumTyInfo = mkHsraEnumTyInfo (Just desc) (mkConstraintInpTy tn) $
-                 mapFromL _eviVal $ map mkConstraintEnumVal cons
-
-    desc = G.Description $
-      "unique or primary key constraints on table " <>> tn
-
-    mkConstraintEnumVal (ConstraintName n) =
-      EnumValInfo (Just "unique or primary key constraint")
-      (G.EnumValue $ G.Name n) False
-
-mkColumnEnumVal :: PGCol -> EnumValInfo
-mkColumnEnumVal (PGCol col) =
-  EnumValInfo (Just "column name") (G.EnumValue $ G.Name col) False
-
-mkUpdColumnTy :: QualifiedTable -> [PGCol] -> EnumTyInfo
-mkUpdColumnTy tn cols = enumTyInfo
-  where
-    enumTyInfo = mkHsraEnumTyInfo (Just desc) (mkUpdColumnInpTy tn) $
-                 mapFromL _eviVal $ map mkColumnEnumVal cols
-
-    desc = G.Description $
-      "update columns of table " <>> tn
-
-mkSelColumnTy :: QualifiedTable -> [PGCol] -> EnumTyInfo
-mkSelColumnTy tn cols = enumTyInfo
-  where
-    enumTyInfo = mkHsraEnumTyInfo (Just desc) (mkSelColumnInpTy tn) $
-                 mapFromL _eviVal $ map mkColumnEnumVal cols
-
-    desc = G.Description $
-      "select columns of table " <>> tn
-
-mkConflictActionTy :: Bool -> EnumTyInfo
-mkConflictActionTy updAllowed =
-  mkHsraEnumTyInfo (Just desc) conflictActionTy $ mapFromL _eviVal $
-  [enumValIgnore] <> bool [] [enumValUpdate] updAllowed
-  where
-    desc = G.Description "conflict action"
-    enumValIgnore = EnumValInfo (Just "ignore the insert on this row")
-                    (G.EnumValue "ignore") False
-    enumValUpdate = EnumValInfo (Just "update the row with the given values")
-                    (G.EnumValue "update") False
-
--- ordByTy :: G.NamedType
--- ordByTy = G.NamedType "order_by"
-
--- ordByEnumTy :: EnumTyInfo
--- ordByEnumTy =
---   mkHsraEnumTyInfo (Just desc) ordByTy $ mapFromL _eviVal $
---   map mkEnumVal enumVals
---   where
---     desc = G.Description "column ordering options"
---     mkEnumVal (n, d) =
---       EnumValInfo (Just d) (G.EnumValue n) False
---     enumVals =
---       [ ( "asc"
---         , "in the ascending order, nulls last"
---         ),
---         ( "desc"
---         , "in the descending order, nulls last"
---         ),
---         ( "asc_nulls_first"
---         , "in the ascending order, nulls first"
---         ),
---         ( "desc_nulls_first"
---         , "in the ascending order, nulls first"
---         )
---       ]
-
-mkTabAggOpOrdByTy :: QualifiedTable -> G.Name -> G.NamedType
-mkTabAggOpOrdByTy tn op =
-  G.NamedType $ qualObjectToName tn <> "_" <> op <> "_order_by"
-
-{-
-input table_<op>_order_by {
-  col1: order_by
-  .     .
-  .     .
-}
--}
-
-mkTabAggOpOrdByInpObjs
-  :: QualifiedTable -> [PGCol] -> [PGCol] -> [InpObjTyInfo]
-mkTabAggOpOrdByInpObjs tn numCols compCols =
-  mapMaybe (mkInpObjTyM numCols) numAggOps
-  <> mapMaybe (mkInpObjTyM compCols) compAggOps
-  where
-
-    mkDesc (G.Name op) = G.Description $ "order by " <> op <> "() on columns of table " <>> tn
-
-    mkInpObjTyM cols op = bool (Just $ mkInpObjTy cols op) Nothing $ null cols
-    mkInpObjTy cols op = mkHsraInpTyInfo (Just $ mkDesc op) (mkTabAggOpOrdByTy tn op) $
-                         fromInpValL $ map mkColInpVal cols
-
-    mkColInpVal c = InpValInfo Nothing (mkColName c) Nothing $ G.toGT
-                    ordByTy
-
-mkTabAggOrdByTy :: QualifiedTable -> G.NamedType
-mkTabAggOrdByTy tn =
-  G.NamedType $ qualObjectToName tn <> "_aggregate_order_by"
-
-{-
-input table_aggregate_order_by {
-count: order_by
-  <op-name>: table_<op-name>_order_by
-}
--}
-
-mkTabAggOrdByInpObj
-  :: QualifiedTable -> [PGCol] -> [PGCol] -> InpObjTyInfo
-mkTabAggOrdByInpObj tn numCols compCols =
-  mkHsraInpTyInfo (Just desc) (mkTabAggOrdByTy tn) $ fromInpValL $
-  numOpOrdBys <> compOpOrdBys <> [countInpVal]
-  where
-    desc = G.Description $
-      "order by aggregate values of table " <>> tn
-
-    numOpOrdBys = bool (map mkInpValInfo numAggOps) [] $ null numCols
-    compOpOrdBys = bool (map mkInpValInfo compAggOps) [] $ null compCols
-    mkInpValInfo op = InpValInfo Nothing op Nothing $ G.toGT $
-                     mkTabAggOpOrdByTy tn op
-
-    countInpVal = InpValInfo Nothing "count" Nothing $ G.toGT ordByTy
-
-mkOrdByTy :: QualifiedTable -> G.NamedType
-mkOrdByTy tn =
-  G.NamedType $ qualObjectToName tn <> "_order_by"
-
-{-
-input table_order_by {
-  col1: order_by
-  col2: order_by
-  .     .
-  .     .
-  coln: order_by
-  obj-rel: <remote-table>_order_by
-}
--}
-
-mkOrdByInpObj
-  :: QualifiedTable -> [SelField] -> (InpObjTyInfo, OrdByCtx)
-mkOrdByInpObj tn selFlds = (inpObjTy, ordByCtx)
-  where
-    inpObjTy =
-      mkHsraInpTyInfo (Just desc) namedTy $ fromInpValL $
-      map mkColOrdBy pgCols <> map mkObjRelOrdBy objRels
-      <> mapMaybe mkArrRelAggOrdBy arrRels
-
-    namedTy = mkOrdByTy tn
-    desc = G.Description $
-      "ordering options when selecting data from " <>> tn
-
-    pgCols = lefts selFlds
-    relFltr ty = flip filter (rights selFlds) $ \(ri, _, _, _, _) ->
-      riType ri == ty
-    objRels = relFltr ObjRel
-    arrRels = relFltr ArrRel
-
-    mkColOrdBy ci = InpValInfo Nothing (mkColName $ pgiName ci) Nothing $
-                    G.toGT ordByTy
-    mkObjRelOrdBy (ri, _, _, _, _) =
-      InpValInfo Nothing (mkRelName $ riName ri) Nothing $
-      G.toGT $ mkOrdByTy $ riRTable ri
-
-    mkArrRelAggOrdBy (ri, isAggAllowed, _, _, _) =
-      let ivi = InpValInfo Nothing (mkAggRelName $ riName ri) Nothing $
-            G.toGT $ mkTabAggOrdByTy $ riRTable ri
-      in bool Nothing (Just ivi) isAggAllowed
-
-    ordByCtx = Map.singleton namedTy $ Map.fromList $
-               colOrdBys <> relOrdBys <> arrRelOrdBys
-    colOrdBys = flip map pgCols $ \ci ->
-                                    ( mkColName $ pgiName ci
-                                    , OBIPGCol ci
-                                    )
-    relOrdBys = flip map objRels $ \(ri, _, fltr, _, _) ->
-                                     ( mkRelName $ riName ri
-                                     , OBIRel ri fltr
-                                     )
-    arrRelOrdBys = flip mapMaybe arrRels $ \(ri, isAggAllowed, fltr, _, _) ->
-                     let obItem = ( mkAggRelName $ riName ri
-                                  , OBIAgg ri fltr
-                                  )
-                     in bool Nothing (Just obItem) isAggAllowed
-
--- newtype RootFlds
---   = RootFlds
---   { _taMutation :: Map.HashMap G.Name (OpCtx, Either ObjFldInfo ObjFldInfo)
---   } deriving (Show, Eq)
-
--- instance Semigroup RootFlds where
---   (RootFlds m1) <> (RootFlds m2)
---     = RootFlds (Map.union m1 m2)
-
--- instance Monoid RootFlds where
---   mempty = RootFlds Map.empty
---   mappend  = (<>)
-
-mkOnConflictTypes
-  :: QualifiedTable -> [ConstraintName] -> [PGCol] -> Bool -> [TypeInfo]
-mkOnConflictTypes tn uniqueOrPrimaryCons cols =
-  bool [] tyInfos
-  where
-    tyInfos = [ TIEnum $ mkConflictActionTy isUpdAllowed
-              , TIEnum $ mkConstriantTy tn uniqueOrPrimaryCons
-              , TIEnum $ mkUpdColumnTy tn cols
-              , TIInpObj $ mkOnConflictInp tn
-              ]
-    isUpdAllowed = not $ null cols
 
 mkGCtxRole'
   :: QualifiedTable
@@ -1348,10 +248,10 @@ mkGCtxRole' tn insPermM selPermM updColsM
         let numCols = (map pgiName . getNumCols) selFlds
             compCols = (map pgiName . getCompCols) selFlds
             objs = [ mkTableAggObj tn
-                   , mkTableAggFldsObj tn numCols compCols
+                   , mkTableAggFldsObj tn (numCols, numAggOps) (compCols, compAggOps)
                    ] <> mkColAggFldsObjs selFlds
-            ordByInps = mkTabAggOrdByInpObj tn numCols compCols
-                        : mkTabAggOpOrdByInpObjs tn numCols compCols
+            ordByInps = mkTabAggOrdByInpObj tn (numCols, numAggOps) (compCols, compAggOps)
+                        : mkTabAggOpOrdByInpObjs tn (numCols, numAggOps) (compCols, compAggOps)
         in (objs, ordByInps)
       _ -> ([], [])
 
@@ -1649,106 +549,6 @@ mkGCtxMapTable tableCache funcCache tabInfo = do
 noFilter :: AnnBoolExpPartialSQL
 noFilter = annBoolExpTrue
 
-checkSchemaConflicts
-  :: (MonadError QErr m)
-  => GCtx -> GCtx -> m ()
-checkSchemaConflicts gCtx remoteCtx = do
-  let typeMap     = _gTypes gCtx -- hasura typemap
-  -- check type conflicts
-  let hTypes      = Map.elems typeMap
-      hTyNames    = map G.unNamedType $ Map.keys typeMap
-      -- get the root names from the remote schema
-      rmQRootName = _otiName $ _gQueryRoot remoteCtx
-      rmMRootName = maybeToList $ _otiName <$> _gMutRoot remoteCtx
-      rmSRootName = maybeToList $ _otiName <$> _gSubRoot remoteCtx
-      rmRootNames = map G.unNamedType (rmQRootName:(rmMRootName ++ rmSRootName))
-  let rmTypes     = Map.filterWithKey
-                    (\k _ -> G.unNamedType k `notElem` builtinTy ++ rmRootNames)
-                    $ _gTypes remoteCtx
-
-      isTyInfoSame ty = any (`tyinfoEq` ty) hTypes
-      -- name is same and structure is not same
-      isSame n ty = G.unNamedType n `elem` hTyNames &&
-                    not (isTyInfoSame ty)
-      conflictedTypes = Map.filterWithKey isSame rmTypes
-      conflictedTyNames = map G.unNamedType $ Map.keys conflictedTypes
-
-  unless (Map.null conflictedTypes) $
-    throw400 RemoteSchemaConflicts $ tyMsg conflictedTyNames
-
-  -- check node conflicts
-  let rmQRoot = _otiFields $ _gQueryRoot remoteCtx
-      rmMRoot = _otiFields <$> _gMutRoot remoteCtx
-      rmRoots = filter (`notElem` builtinNodes ++ rmRootNames) . Map.keys <$>
-                mergeMaybeMaps (Just rmQRoot) rmMRoot
-      hQR     = _otiFields <$>
-                join (getObjTyM <$> Map.lookup hQRName typeMap)
-      hMR     = _otiFields <$>
-                join (getObjTyM <$> Map.lookup hMRName typeMap)
-      hRoots  = Map.keys <$> mergeMaybeMaps hQR hMR
-
-  case (rmRoots, hRoots) of
-    (Just rmR, Just hR) -> do
-      let conflictedNodes = filter (`elem` hR) rmR
-      unless (null conflictedNodes) $
-        throw400 RemoteSchemaConflicts $ nodesMsg conflictedNodes
-    _ -> return ()
-
-  where
-    tyinfoEq a b = case (a, b) of
-      (TIScalar t1, TIScalar t2) -> typeEq t1 t2
-      (TIObj t1, TIObj t2)       -> typeEq t1 t2
-      (TIEnum t1, TIEnum t2)     -> typeEq t1 t2
-      (TIInpObj t1, TIInpObj t2) -> typeEq t1 t2
-      _                          -> False
-
-    hQRName = G.NamedType "query_root"
-    hMRName = G.NamedType "mutation_root"
-    tyMsg ty = "types: [ " <> namesToTxt ty <>
-               " ] have mismatch with current graphql schema. HINT: Types must be same."
-    nodesMsg n = "top-level nodes: [ " <> namesToTxt n <>
-                 " ] already exist in current graphql schema. HINT: Top-level nodes can't be same."
-    namesToTxt = T.intercalate ", " . map G.unName
-    builtinNodes = ["__type", "__schema", "__typename"]
-    builtinTy = [ "__Directive"
-                , "__DirectiveLocation"
-                , "__EnumValue"
-                , "__Field"
-                , "__InputValue"
-                , "__Schema"
-                , "__Type"
-                , "__TypeKind"
-                , "Int"
-                , "Float"
-                , "String"
-                , "Boolean"
-                , "ID"
-                ]
-
-checkConflictingNode
-  :: (MonadError QErr m)
-  => GCtx
-  -> G.Name
-  -> m ()
-checkConflictingNode gCtx node = do
-  let typeMap = _gTypes gCtx
-      hQR     = _otiFields <$>
-                join (getObjTyM <$> Map.lookup hQRName typeMap)
-      hMR     = _otiFields <$>
-                join (getObjTyM <$> Map.lookup hMRName typeMap)
-      hRoots  = Map.keys <$> mergeMaybeMaps hQR hMR
-  case hRoots of
-    Just hR ->
-      when (node `elem` hR) $
-        throw400 RemoteSchemaConflicts msg
-    _ -> return ()
-  where
-    hQRName = G.NamedType "query_root"
-    hMRName = G.NamedType "mutation_root"
-    msg = "node " <> G.unName node <>
-          " already exists in current graphql schema"
-
-
 mkGCtxMap
   :: (MonadError QErr m)
   => TableCache -> FunctionCache -> m GCtxMap
@@ -1775,18 +575,6 @@ getGCtx :: (CacheRM m) => RoleName -> GCtxMap -> m GCtx
 getGCtx rn ctxMap = do
   sc <- askSchemaCache
   return $ fromMaybe (scDefaultRemoteGCtx sc) $ Map.lookup rn ctxMap
-
-mergeMaybeMaps
-  :: (Eq k, Hashable k)
-  => Maybe (Map.HashMap k v)
-  -> Maybe (Map.HashMap k v)
-  -> Maybe (Map.HashMap k v)
-mergeMaybeMaps m1 m2 = case (m1, m2) of
-  (Nothing, Nothing)   -> Nothing
-  (Just m1', Nothing)  -> Just m1'
-  (Nothing, Just m2')  -> Just m2'
-  (Just m1', Just m2') -> Just $ Map.union m1' m2'
-
 
 -- pretty print GCtx
 ppGCtx :: GCtx -> String
@@ -1840,218 +628,6 @@ instance Monoid RootFlds where
   mempty = RootFlds Map.empty
   mappend  = (<>)
 
-fromInpValL :: [InpValInfo] -> Map.HashMap G.Name InpValInfo
-fromInpValL = mapFromL _iviName
-
-mkCompExpName :: PGColType -> G.Name
-mkCompExpName pgColTy =
-  G.Name $ T.pack (show pgColTy) <> "_comparison_exp"
-
-mkCompExpTy :: PGColType -> G.NamedType
-mkCompExpTy =
-  G.NamedType . mkCompExpName
-
-mkCastExpName :: PGColType -> G.Name
-mkCastExpName pgColTy = G.Name $ T.pack (show pgColTy) <> "_cast_exp"
-
-mkCastExpTy :: PGColType -> G.NamedType
-mkCastExpTy = G.NamedType . mkCastExpName
-
--- TODO(shahidhk) this should ideally be st_d_within_geometry
-{-
-input st_d_within_input {
-  distance: Float!
-  from: geometry!
-}
--}
-stDWithinGeometryInpTy :: G.NamedType
-stDWithinGeometryInpTy = G.NamedType "st_d_within_input"
-
-{-
-input st_d_within_geography_input {
-  distance: Float!
-  from: geography!
-  use_spheroid: Bool!
-}
--}
-stDWithinGeographyInpTy :: G.NamedType
-stDWithinGeographyInpTy = G.NamedType "st_d_within_geography_input"
-
-
---- | make compare expression input type
-mkCompExpInp :: PGColType -> InpObjTyInfo
-mkCompExpInp colTy =
-  InpObjTyInfo (Just tyDesc) (mkCompExpTy colTy) (fromInpValL $ concat
-  [ map (mk colScalarTy) typedOps
-  , map (mk $ G.toLT $ G.toNT colScalarTy) listOps
-  , bool [] (map (mk $ mkScalarTy PGText) stringOps) isStringTy
-  , bool [] (map jsonbOpToInpVal jsonbOps) isJsonbTy
-  , bool [] (stDWithinGeoOpInpVal stDWithinGeometryInpTy :
-             map geoOpToInpVal (geoOps ++ geomOps)) isGeometryType
-  , bool [] (stDWithinGeoOpInpVal stDWithinGeographyInpTy :
-             map geoOpToInpVal geoOps) isGeographyType
-  , [InpValInfo Nothing "_is_null" Nothing $ G.TypeNamed (G.Nullability True) $ G.NamedType "Boolean"]
-  , maybeToList castOpInputValue
-  ]) TLHasuraType
-  where
-    tyDesc = mconcat
-      [ "expression to compare columns of type "
-      , G.Description (T.pack $ show colTy)
-      , ". All fields are combined with logical 'AND'."
-      ]
-    isStringTy = case colTy of
-      PGVarchar -> True
-      PGText    -> True
-      _         -> False
-    mk t n = InpValInfo Nothing n Nothing $ G.toGT t
-    colScalarTy = mkScalarTy colTy
-    -- colScalarListTy = GA.GTList colGTy
-    typedOps =
-       ["_eq", "_neq", "_gt", "_lt", "_gte", "_lte"]
-    listOps =
-      [ "_in", "_nin" ]
-    -- TODO
-    -- columnOps =
-    --   [ "_ceq", "_cneq", "_cgt", "_clt", "_cgte", "_clte"]
-    stringOps =
-      [ "_like", "_nlike", "_ilike", "_nilike"
-      , "_similar", "_nsimilar"
-      ]
-
-    isJsonbTy = case colTy of
-      PGJSONB -> True
-      _       -> False
-    jsonbOpToInpVal (op, ty, desc) = InpValInfo (Just desc) op Nothing ty
-    jsonbOps =
-      [ ( "_contains"
-        , G.toGT $ mkScalarTy PGJSONB
-        , "does the column contain the given json value at the top level"
-        )
-      , ( "_contained_in"
-        , G.toGT $ mkScalarTy PGJSONB
-        , "is the column contained in the given json value"
-        )
-      , ( "_has_key"
-        , G.toGT $ mkScalarTy PGText
-        , "does the string exist as a top-level key in the column"
-        )
-      , ( "_has_keys_any"
-        , G.toGT $ G.toLT $ G.toNT $ mkScalarTy PGText
-        , "do any of these strings exist as top-level keys in the column"
-        )
-      , ( "_has_keys_all"
-        , G.toGT $ G.toLT $ G.toNT $ mkScalarTy PGText
-        , "do all of these strings exist as top-level keys in the column"
-        )
-      ]
-
-    castOpInputValue =
-      -- currently, only geometry/geography types support casting
-      guard (isGeoType colTy) $>
-        InpValInfo Nothing "_cast" Nothing (G.toGT $ mkCastExpTy colTy)
-
-    stDWithinGeoOpInpVal ty =
-      InpValInfo (Just stDWithinGeoDesc) "_st_d_within" Nothing $ G.toGT ty
-    stDWithinGeoDesc =
-      "is the column within a distance from a " <> colTyDesc <> " value"
-
-    -- Geometry related ops
-    isGeometryType = case colTy of
-      PGGeometry -> True
-      _          -> False
-
-    -- Geography related ops
-    isGeographyType = case colTy of
-      PGGeography -> True
-      _           -> False
-
-    geoOpToInpVal (op, desc) =
-      InpValInfo (Just desc) op Nothing $ G.toGT $ mkScalarTy colTy
-
-    colTyDesc = G.Description $ T.pack $ show colTy
-
-    -- operators applicable only to geometry types
-    geomOps :: [(G.Name, G.Description)]
-    geomOps =
-      [
-        ( "_st_contains"
-        , "does the column contain the given geometry value"
-        )
-      , ( "_st_crosses"
-        , "does the column crosses the given geometry value"
-        )
-      , ( "_st_equals"
-        , "is the column equal to given geometry value. Directionality is ignored"
-        )
-      , ( "_st_overlaps"
-        , "does the column 'spatially overlap' (intersect but not completely contain) the given geometry value"
-        )
-      , ( "_st_touches"
-        , "does the column have atleast one point in common with the given geometry value"
-        )
-      , ( "_st_within"
-        , "is the column contained in the given geometry value"
-        )
-      ]
-
-    -- operators applicable to geometry and geography types
-    geoOps =
-      [
-        ( "_st_intersects"
-        , "does the column spatially intersect the given " <> colTyDesc <> " value"
-        )
-      ]
-
--- | Makes an input type declaration for the @_cast@ field of a comparison expression.
--- (Currently only used for casting between geometry and geography types.)
-mkCastExpressionInputType :: PGColType -> [PGColType] -> InpObjTyInfo
-mkCastExpressionInputType sourceType targetTypes =
-  mkHsraInpTyInfo (Just description) (mkCastExpTy sourceType) (fromInpValL targetFields)
-  where
-    description = mconcat
-      [ "Expression to compare the result of casting a column of type "
-      , G.Description (T.pack $ show sourceType)
-      , ". Multiple cast targets are combined with logical 'AND'."
-      ]
-    targetFields = map targetField targetTypes
-    targetField targetType = InpValInfo
-      Nothing
-      (G.Name . T.pack $ show targetType)
-      Nothing
-      (G.toGT $ mkCompExpTy targetType)
-
-ordByTy :: G.NamedType
-ordByTy = G.NamedType "order_by"
-
-ordByEnumTy :: EnumTyInfo
-ordByEnumTy =
-  mkHsraEnumTyInfo (Just desc) ordByTy $ mapFromL _eviVal $
-  map mkEnumVal enumVals
-  where
-    desc = G.Description "column ordering options"
-    mkEnumVal (n, d) =
-      EnumValInfo (Just d) (G.EnumValue n) False
-    enumVals =
-      [ ( "asc"
-        , "in the ascending order, nulls last"
-        ),
-        ( "asc_nulls_last"
-        , "in the ascending order, nulls last"
-        ),
-        ( "asc_nulls_first"
-        , "in the ascending order, nulls first"
-        ),
-        ( "desc"
-        , "in the descending order, nulls first"
-        ),
-        ( "desc_nulls_first"
-        , "in the descending order, nulls first"
-        ),
-        ( "desc_nulls_last"
-        , "in the descending order, nulls last"
-        )
-      ]
-
 mkGCtx :: TyAgg -> RootFlds -> InsCtxMap -> GCtx
 mkGCtx tyAgg (RootFlds flds) insCtxMap =
   let queryRoot = mkQueryRootTyInfo qFlds
@@ -2090,26 +666,4 @@ mkGCtx tyAgg (RootFlds flds) insCtxMap =
         else colTys
     allScalarTypes = allComparableTypes <> scalars
 
-    wiredInGeoInputTypes =
-      guard anyGeoTypes *> map TIInpObj
-        [ stDWithinGeometryInputType
-        , stDWithinGeographyInputType
-        , castGeometryInputType
-        , castGeographyInputType
-        ]
-
-    stDWithinGeometryInputType =
-      mkHsraInpTyInfo Nothing stDWithinGeometryInpTy $ fromInpValL
-      [ InpValInfo Nothing "from" Nothing $ G.toGT $ G.toNT $ mkScalarTy PGGeometry
-      , InpValInfo Nothing "distance" Nothing $ G.toNT $ mkScalarTy PGFloat
-      ]
-    stDWithinGeographyInputType =
-      mkHsraInpTyInfo Nothing stDWithinGeographyInpTy $ fromInpValL
-      [ InpValInfo Nothing "from" Nothing $ G.toGT $ G.toNT $ mkScalarTy PGGeography
-      , InpValInfo Nothing "distance" Nothing $ G.toNT $ mkScalarTy PGFloat
-      , InpValInfo
-        Nothing "use_spheroid" (Just $ G.VCBoolean True) $ G.toGT $ mkScalarTy PGBoolean
-      ]
-
-    castGeometryInputType = mkCastExpressionInputType PGGeometry [PGGeography]
-    castGeographyInputType = mkCastExpressionInputType PGGeography [PGGeometry]
+    wiredInGeoInputTypes = guard anyGeoTypes *> map TIInpObj geoInputTypes
